Added iter_*_mut variations to HalfEdgeGraph
I also had to add a new *FnData type, to try to mimic *FnMut but
without exclusive access to the graph.

It should be doable with *FnMut, but would be tricky to implmenent
without access to SlotMap's internals.

The problem is that to iterate over all values we need to borrow the map,
and in extension the graph. This means that we can't give out mutable
references to the graph (as we would using *FnMut), and instead settle
with the references that slotmaps gives us, which is the key + value pair